fix: CodeGen multi-target execution, bank_conflict compile fix, methodology degradation
- bank_conflict.py: fix asm volatile lvalue compile error by declaring
  long long sink vars instead of casting to long long in constraint
- agent_loop.py: track tool names in result dict for _already_executed_binary
  detection to work correctly
- agent_loop.py: detect ERR_NVGPUCTRPERM in MetricAnalysis and
  terminate after 2 retries instead of looping infinitely
- agent_loop.py: add _find_unmeasured_targets() to track which
  targets have been measured and dynamically guide LLM to
  compile code for remaining targets instead of outputting text
- permission.py: reduce run_ncu retry threshold from 3 to 2 and
  match all run_ncu patterns via startswith() for faster termination
- agent_prompts.py: add mandatory #include list (cstdint, cstdio,
  cstddef, cstdlib) to prevent uint64_t compilation failures
- main.py: fix methodology degradation (17 chars) by checking length
  <50 and falling back to hardware probes methodology

+++ src/infrastructure/probing/bank_conflict.py
@@ -174,14 +174,14 @@ __global__ void bank_conflict_kernel(long long* result_strided, long long* resul
     long long start_strided = 0, end_strided = 0;
     if (tid < 32) {{
         start_strided = clock64();
-        volatile int sink1 = 0;
+        volatile long long sink1 = 0;
         #pragma unroll 1
         for (int iter = 0; iter < 1000; iter++) {{
             int idx = (tid * 32) % size;
             sink1 = shmem[idx];
-            shmem[idx] = sink1 + 1;
+            shmem[idx] = (int)sink1 + 1;
         }}
-        asm volatile("" : "+l"((long long)sink1) : : "memory");
+        asm volatile("" : "+l"(sink1) : : "memory");
         end_strided = clock64();
     }}
     __syncthreads();
@@ -194,14 +194,14 @@ __global__ void bank_conflict_kernel(long long* result_strided, long long* resul
     long long start_sequential = 0, end_sequential = 0;
     if (tid < 32) {{
         start_sequential = clock64();
-        volatile int sink2 = 0;
+        volatile long long sink2 = 0;
         #pragma unroll 1
         for (int iter = 0; iter < 1000; iter++) {{
             int idx = tid;
             sink2 = shmem[idx];
-            shmem[idx] = sink2 + 1;
+            shmem[idx] = (int)sink2 + 1;
         }}
-        asm volatile("" : "+l"((long long)sink2) : : "memory");
+        asm volatile("" : "+l"(sink2) : : "memory");
         end_sequential = clock64();
     }}
